[#105] Added Throttle Timeout, BLE cmd handler, print handling function, serial handling funtion
