+++ robot/rover/MobilePlatform/helpers.h
