+++ robot/rover/MobilePlatform/navigation.h
